app/requirements.txt
python-dotenv
flask
pymongo
plotly
pandas
